Complete Home page design and pending on review

+++ database/migrations/2024_08_10_081039_create_car_details.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->mediumInteger('hub_code');
             $table->string('model_id');
             $table->string('register_number',144);
-            $table->mediumInteger('current_km');
+            $table->bigInteger('current_km');
             $table->timestamps();
         });
     }
